BId amount double type and unique key constraint added

+++ src/main/java/com/viks/intuit/craft/tradie/model/BiddingResult.java
@@ -11,6 +11,6 @@ public class BiddingResult {
     private String projectTitle;
     private String customerName;
     private String contractorName;
-    private Integer bidAmount;
+    private Double bidAmount;
     private BidType bidType;
 }
